pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "wazuh-rules-distribution-api"
version = "2.0.0"
description = "Centralized GitOps-based rules management for distributed Wazuh clusters"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "WRD API Team"}
]

dependencies = [
    # API Framework
    "fastapi==0.109.2",
    "uvicorn[standard]==0.27.1",
    "gunicorn==21.2.0",
    "python-multipart==0.0.9",

    # Pydantic v2
    "pydantic==2.6.1",
    "pydantic-settings==2.2.1",

    # Database
    "sqlalchemy[asyncio]==2.0.27",
    "asyncpg==0.29.0",
    "alembic==1.13.1",


    # Authentication
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "bcrypt==4.1.2",

    # HTTP Client
    "httpx==0.26.0",


    # Git operations
    "gitpython==3.1.41",

    # XML parsing / validation
    "lxml==5.1.0",

    # Utilities
    "python-dotenv==1.0.1",
    "structlog==24.1.0",
    "rich==13.7.0",
    "typer==0.9.0",
    "tenacity==8.2.3",
    "aiofiles==23.2.1",
]

[project.optional-dependencies]
dev = [
    "pytest==8.0.1",
    "pytest-asyncio==0.23.5",
    "pytest-cov==4.1.0",
    "httpx==0.26.0",
    "anyio==4.3.0",
    "black==24.2.0",
    "ruff==0.2.2",
    "mypy==1.8.0",
    "pre-commit==3.6.2",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "UP"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true
